Breast_Cancer_Analysis_Project/requirements.txt
streamlit
pandas
numpy 
sklearn.preprocessing
hpelm